+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+* 2024/02/04
+  * view model compositor 支持通过前缀访问子vm的属性。
+
 * 2024/02/21
   * 完善 MVVM 绑定的 can exec。（感谢兆坤提供补丁）
   * command binding 添加 is_continue 属性，主要解决控件同时绑定value_changed 和 value_changed_by_ui命令时，有个会不执行的问题。（感谢兆坤提供补丁）